Move GCHandle to shared
